# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export type SendAnthropicMessageBodyModel =
  (typeof SendAnthropicMessageBodyModel)[keyof typeof SendAnthropicMessageBodyModel];

export const SendAnthropicMessageBodyModel = {
  sonnet: "sonnet",
  opus: "opus",
  haiku: "haiku",
} as const;

export interface SendAnthropicMessageBody {
  content: string;
  model?: SendAnthropicMessageBodyModel;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface ResearchJob {
  id: number;
  title: string;
  sourceType: string;
  status: string;
  targets: string;
  summary?: string | null;
  apifyRunId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ResearchResult {
  id: number;
  jobId: number;
  url?: string | null;
  title?: string | null;
  content: string;
  createdAt: string;
}

export type CreateResearchJobBodySourceType =
  (typeof CreateResearchJobBodySourceType)[keyof typeof CreateResearchJobBodySourceType];

export const CreateResearchJobBodySourceType = {
  website: "website",
  competitor: "competitor",
  reviews: "reviews",
  search: "search",
  social: "social",
  custom: "custom",
} as const;

export interface CreateResearchJobBody {
  title: string;
  sourceType: CreateResearchJobBodySourceType;
  targets: string;
  scrapeTemplate?: string;
}

export interface ContentAsset {
  id: number;
  title: string;
  type: string;
  content: string;
  platform?: string | null;
  tone?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContentAssetBody {
  title: string;
  type: string;
  content: string;
  platform?: string;
  tone?: string;
}

export type GenerateContentBodyType =
  (typeof GenerateContentBodyType)[keyof typeof GenerateContentBodyType];

export const GenerateContentBodyType = {
  ad_copy: "ad_copy",
  email: "email",
  landing_page: "landing_page",
  social: "social",
  headline: "headline",
  hook: "hook",
  offer: "offer",
  cta: "cta",
  blog_outline: "blog_outline",
} as const;

export type GenerateContentBodyModel =
  (typeof GenerateContentBodyModel)[keyof typeof GenerateContentBodyModel];

export const GenerateContentBodyModel = {
  sonnet: "sonnet",
  opus: "opus",
  haiku: "haiku",
} as const;

export interface GenerateContentBody {
  type: GenerateContentBodyType;
  platform?: string;
  tone?: string;
  audience?: string;
  context: string;
  model?: GenerateContentBodyModel;
  variants?: number;
}

export interface Campaign {
  id: number;
  title: string;
  description?: string | null;
  status: string;
  objective?: string | null;
  audience?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCampaignBody {
  title: string;
  description?: string;
  objective?: string;
  audience?: string;
}

export interface UpdateCampaignBody {
  title?: string;
  description?: string;
  status?: string;
  objective?: string;
  audience?: string;
}

export interface KnowledgeItem {
  id: number;
  title: string;
  type: string;
  content: string;
  tags?: string | null;
  url?: string | null;
  includeInContext: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateKnowledgeItemBodyType =
  (typeof CreateKnowledgeItemBodyType)[keyof typeof CreateKnowledgeItemBodyType];

export const CreateKnowledgeItemBodyType = {
  note: "note",
  url: "url",
  file: "file",
  scraped: "scraped",
} as const;

export interface CreateKnowledgeItemBody {
  title: string;
  type: CreateKnowledgeItemBodyType;
  content: string;
  tags?: string;
  url?: string;
  includeInContext?: boolean;
}

export interface Automation {
  id: number;
  title: string;
  description?: string | null;
  trigger: string;
  action: string;
  enabled: boolean;
  lastRunAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAutomationBody {
  title: string;
  description?: string;
  trigger: string;
  action: string;
}

export interface UpdateAutomationBody {
  title?: string;
  description?: string;
  trigger?: string;
  action?: string;
  enabled?: boolean;
}

export interface AutomationRun {
  id: number;
  automationId: number;
  status: string;
  output?: string | null;
  createdAt: string;
}

export interface Project {
  id: number;
  title: string;
  description?: string | null;
  industry?: string | null;
  useCase?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  title: string;
  description?: string;
  industry?: string;
  useCase?: string;
}

export interface ActivityItem {
  type: string;
  title: string;
  createdAt: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalCampaigns: number;
  totalResearchJobs: number;
  totalContentAssets: number;
  totalKnowledgeItems: number;
  totalAutomations: number;
  activeAutomations: number;
  recentActivity: ActivityItem[];
}

export interface ApiError {
  error: string;
}