# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ATREYU API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content/assets:
    get:
      operationId: listContentAssets
      tags: [content]
      summary: List all content assets
      responses:
        "200":
          description: List of content assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentAsset"
    post:
      operationId: createContentAsset
      tags: [content]
      summary: Create a content asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentAssetBody"
      responses:
        "201":
          description: Created content asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset"
  /content/assets/{id}:
    get:
      operationId: getContentAsset
      tags: [content]
      summary: Get a content asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset"
    delete:
      operationId: deleteContentAsset
      tags: [content]
      summary: Delete a content asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content/generate:
    post:
      operationId: generateContent
      tags: [content]
      summary: Generate content using AI (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentBody"
      responses:
        "200":
          description: SSE stream of generated content
          content:
            text/event-stream: {}
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    put:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [knowledge]
      summary: List all knowledge items
      responses:
        "200":
          description: List of knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      operationId: createKnowledgeItem
      tags: [knowledge]
      summary: Create a knowledge item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeItemBody"
      responses:
        "201":
          description: Created knowledge item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    get:
      operationId: getKnowledgeItem
      tags: [knowledge]
      summary: Get a knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Knowledge item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
    delete:
      operationId: deleteKnowledgeItem
      tags: [knowledge]
      summary: Delete a knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List all automations
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create an automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    get:
      operationId: getAutomation
      tags: [automations]
      summary: Get an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automations/{id}/toggle:
    post:
      operationId: toggleAutomation
      tags: [automations]
      summary: Toggle automation enabled state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}/run:
    post:
      operationId: runAutomation
      tags: [automations]
      summary: Manually trigger an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
...
components:
  schemas:
...
    ContentAsset:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        content:
          type: string
        platform:
          type: string
          nullable: true
        tone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - content
        - createdAt
        - updatedAt
    CreateContentAssetBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        content:
          type: string
        platform:
          type: string
        tone:
          type: string
      required:
        - title
        - type
        - content
    GenerateContentBody:
      type: object
      properties:
        type:
          type: string
          enum: [ad_copy, email, landing_page, social, headline, hook, offer, cta, blog_outline]
        platform:
          type: string
        tone:
          type: string
        audience:
          type: string
        context:
          type: string
        model:
          type: string
          enum: [sonnet, opus, haiku]
        variants:
          type: integer
      required:
        - type
        - context
    Campaign:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        objective:
          type: string
          nullable: true
        audience:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
    CreateCampaignBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        objective:
          type: string
        audience:
          type: string
      required:
        - title
    UpdateCampaignBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        objective:
          type: string
        audience:
          type: string
    KnowledgeItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        content:
          type: string
        tags:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        includeInContext:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - content
        - includeInContext
        - createdAt
        - updatedAt
    CreateKnowledgeItemBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [note, url, file, scraped]
        content:
          type: string
        tags:
          type: string
        url:
          type: string
        includeInContext:
          type: boolean
      required:
        - title
        - type
        - content
    Automation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        trigger:
          type: string
        action:
          type: string
        enabled:
          type: boolean
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - trigger
        - action
        - enabled
        - createdAt
        - updatedAt
    CreateAutomationBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        trigger:
          type: string
        action:
          type: string
      required:
        - title
        - trigger
        - action
    UpdateAutomationBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        trigger:
          type: string
        action:
          type: string
        enabled:
          type: boolean
    AutomationRun:
      type: object
      properties:
        id:
          type: integer
        automationId:
          type: integer
        status:
          type: string
        output:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - automationId
        - status
        - createdAt
...