# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Automation {
  id: number;
  title: string;
  description?: string | null;
  trigger: string;
  action: string;
  enabled: boolean;
  lastRunAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}