# lib/api-zod/src/generated/types/createContentAssetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContentAssetBody {
  title: string;
  type: string;
  content: string;
  platform?: string;
  tone?: string;
}