# lib/api-zod/src/generated/types/createKnowledgeItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateKnowledgeItemBodyType } from "./createKnowledgeItemBodyType";

export interface CreateKnowledgeItemBody {
  title: string;
  type: CreateKnowledgeItemBodyType;
  content: string;
  tags?: string;
  url?: string;
  includeInContext?: boolean;
}