# lib/api-zod/src/generated/types/updateAutomationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAutomationBody {
  title?: string;
  description?: string;
  trigger?: string;
  action?: string;
  enabled?: boolean;
}