# lib/api-zod/src/generated/types/updateCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATREYU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCampaignBody {
  title?: string;
  description?: string;
  status?: string;
  objective?: string;
  audience?: string;
}